Lab boundingbox raycasting

+++ WRInteractiveGraphicsS24/IndexBuffer.h
@@ -16,7 +16,7 @@ public:
 	inline void Select() { glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, iboId); }
 	inline void Deselect() { glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0); }
 	inline const std::vector<unsigned short>& GetIndexData() { return indexData; }
-	inline unsigned int GetNumberOfIndexes() const { return indexData.size(); }
+	inline unsigned int GetNumberOfIndexes() const { return (unsigned int)indexData.size(); }
 	void AddIndexData(unsigned int count, ...);
 	void AddIndexData(unsigned short index);
 	void StaticAllocate();
